email-search/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
sentence-transformers==3.1.1
chromadb==0.5.18
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
httpx==0.27.2
rank-bm25==0.2.2
rich==13.9.2
click==8.1.7
python-multipart==0.0.12
itsdangerous==2.2.0
